Only allow staff to edit divisions and policies

+++ app/controllers/divisions_controller.rb
@@ -94,6 +94,7 @@ class DivisionsController < ApplicationController
 
   def edit
     @division = Division.in_house(params[:house] || "representatives").find_by!(date: params[:date], number: params[:number])
+    authorize @division
   end
 
   def history
@@ -102,6 +103,7 @@ class DivisionsController < ApplicationController
 
   def update
     @division = Division.in_house(params[:house] || "representatives").find_by!(date: params[:date], number: params[:number])
+    authorize @division
 
     wiki_motion = @division.build_wiki_motion(params[:newtitle], params[:newdescription], current_user)
 
@@ -112,9 +114,11 @@ class DivisionsController < ApplicationController
     end
   end
 
+  # TODO: Move this to a policy_division controller
   def create_policy_division
     @division = Division.in_house(params[:house]).find_by!(date: params[:date], number: params[:number])
     @policy_division = @division.policy_divisions.new(policy_division_params)
+    authorize @policy_division, :create?
 
     if @policy_division.save
       # TODO: Just point to the object when the path helper has been refactored
@@ -126,9 +130,11 @@ class DivisionsController < ApplicationController
     end
   end
 
+  # TODO: Move this to a policy_division controller
   def update_policy_division
     division = Division.in_house(params[:house]).find_by!(date: params[:date], number: params[:number])
     policy_division = PolicyDivision.find_by!(division: division, policy: params[:policy_id])
+    authorize policy_division, :update?
 
     if policy_division.update(policy_division_params)
       flash[:notice] = "Updated policy connection"
@@ -140,9 +146,11 @@ class DivisionsController < ApplicationController
     redirect_to division_policies_path(house: division.house, date: division.date, number: division.number)
   end
 
+  # TODO: Move this to a policy_division controller
   def destroy_policy_division
     division = Division.in_house(params[:house]).find_by!(date: params[:date], number: params[:number])
     policy_division = PolicyDivision.find_by!(division: division, policy: params[:policy_id])
+    authorize policy_division, :destroy?
 
     if policy_division.destroy
       flash[:notice] = "Removed policy connection"
